+++ integrations/sentry/clients/sentry.py
@@ -6,12 +6,12 @@ from integration import SentryResourceConfig
 from loguru import logger
 from port_ocean.context.event import event
 from port_ocean.helpers.async_client import OceanAsyncClient
+from port_ocean.helpers.retry import RetryConfig
 from port_ocean.utils.async_iterators import stream_async_iterators_tasks
 from port_ocean.utils.cache import cache_iterator_result
 
 from .exceptions import IgnoredError, ResourceNotFoundError
 from .rate_limiter import SentryRateLimiter
-from .retry_transport import SentryRetryTransport
 
 PAGE_SIZE = 100
 
@@ -42,7 +42,11 @@ class SentryClient:
         self.base_headers = {"Authorization": "Bearer " + f"{self.auth_token}"}
         self.api_url = f"{self.sentry_base_url}/api/0"
         self.organization = sentry_organization
-        self._client = OceanAsyncClient(transport_class=SentryRetryTransport)
+        retry_config = RetryConfig(
+            retry_after_headers=["X-Sentry-Rate-Limit-Reset", "Retry-After"],
+            max_attempts=5,
+        )
+        self._client = OceanAsyncClient(retry_config=retry_config)
         self.selector = cast(SentryResourceConfig, event.resource_config).selector
         self._rate_limiter = SentryRateLimiter()